refactor: remove residual coordinator and facade indirection
